fix) community delete기능시, findByuserId로 찾던 기능을 findBycommunityId로 변경

+++ travel/src/main/java/com/zerobase/travel/communities/service/CommunityService.java
@@ -64,8 +64,13 @@ public class CommunityService {
 
     public void deletePost(long communityId, String userId) {
 
-        communityRepository.findByUserId(userId).orElseThrow(()
-            -> new CustomException(ErrorCode.USER_UNAUTHORIZED_REQUEST));
+        CommunityEntity communityEntity = communityRepository.findById(
+            communityId).orElseThrow(()
+            -> new CustomException(ErrorCode.COMMUNITY_NON_EXISTENT));
+
+        if(!communityEntity.getUserId().equals(userId)) {
+            throw new CustomException(ErrorCode.USER_UNAUTHORIZED_REQUEST);
+        }
 
         communityRepository.deleteByCommunityId(communityId);
     }
